rename: ga4->GA4 , fix: _app.tsx
